requirements.txt
torch>=1.9.0
torchvision>=0.10.0
PyQt5>=5.15.4
matplotlib>=3.4.3
numpy>=1.21.2
pillow>=8.3.1